fix(CRED-310): Allow decimal points for interest rate

+++ custom/crediblex/portfolio/loc/src/main/java/com/crediblex/fineract/portfolio/loc/data/LineOfCreditRequest.java
@@ -104,7 +104,7 @@ public class LineOfCreditRequest {
 
     @Schema(example = "3", description = "Annual interest rate")
     @NotNull(message = "Annual interest rate is required")
-    private Integer annualInterestRate;
+    private Float annualInterestRate;
 
     @Schema(example = "365", description = "Tenor in days")
     @NotNull(message = "Tenor days is required")
